Retrieve printerdata directly from db
Due to network changes we can't rely on ip restrictions
to access the API that ran on p.foreningenbs.no. That API is
also scheduled to be removed, and since the DB no longer runs
on that machine we can access it directly.

The Printer class and PrinterConfig class are more or less just
copied from the old API.